@@ -24,7 +24,6 @@
     "*.ts": "lint",
     "js/services/*.ts": "lint:docs"
   },
-  "main": "publish/js/ui-core.js",
   "files": [
     "publish/"
   ],
@@ -32,39 +31,38 @@
     "@types/jquery": "3.2.11",
     "@types/lodash": "4.14.101",
     "@types/mockery": "1.4.29",
-    "@types/node": "8.0.30",
+    "@types/node": "12.7.2",
     "@types/react": "16.0.2",
-    "@types/sinon": "2.3.3",
+    "@types/sinon": "7.0.13",
     "@types/socket.io-client": "1.4.30",
     "ava": "0.22.0",
-    "awesome-typescript-loader": "3.2.3",
+    "awesome-typescript-loader": "^5.0.0-1",
     "browser-env": "3.1.2",
     "chokidar-cli": "1.2.2",
     "clean-webpack-plugin": "0.1.19",
     "del": "4.1.1",
     "form-data": "2.2.0",
-    "husky": "0.14.3",
+    "husky": "3.0.4",
     "lint-staged": "4.0.1",
-    "mock-socket": "7.0.0",
+    "mock-socket": "9.0.0",
     "mockery": "2.1.0",
     "nyc": "14.1.0",
     "script-loader": "0.7.0",
     "sinon": "3.2.1",
-    "tap-xunit": "2.3.0",
+    "tap-xunit": "2.4.1",
     "ts-loader": "4.4.2",
     "tslint": "5.8.0",
     "tslint-config-airbnb": "5.9.2",
     "tslint-loader": "3.6.0",
     "typedoc": "0.14.2",
-    "typescript": "2.4.2",
+    "typescript": "2.9.2",
     "uglifyjs-webpack-plugin": "0.4.6",
     "webpack": "4.26.1",
     "webpack-bundle-analyzer": "3.3.2",
-    "webpack-cli": "3.1.2",
+    "webpack-cli": "3.3.7",
     "xhr-mock": "1.9.0"
   },
   "dependencies": {
-    "compression-webpack-plugin": "2.0.0",
     "jquery": "3.4.1",
     "license-webpack-plugin": "1.1.1",
     "loglevel": "1.4.1",
@@ -72,7 +70,7 @@
     "numbro": "1.11.0",
     "react": "15.6.1",
     "react-dom": "15.6.1",
-    "socket.io": "2.0.3",
+    "socket.io": "2.2.0",
     "socket.io-client": "2.0.3",
     "webpack-dashboard": "3.0.5",
     "write-file": "1.0.0"
